Former-commit-id: 3ba988c4d3bc788d18dac82417166222db4c2a9b

+++ pybinsim/convolver.py
@@ -113,8 +113,8 @@ class ConvolverTorch(object):
         """
 
         left, right = current_filter.getFilterFD()
-        self.TF_left_blocked = torch.as_tensor(left, device=self.torch_device)
-        self.TF_right_blocked = torch.as_tensor(right, device=self.torch_device)
+        self.TF_left_blocked = torch.as_tensor(left, dtype=torch.complex64, device=self.torch_device)
+        self.TF_right_blocked = torch.as_tensor(right, dtype=torch.complex64, device=self.torch_device)
 
 
         # Interpolation means cross fading the output blocks (linear interpolation)
